feat: 인증 구조 개선 및 login 리팩토링
- Firebase 초기화 문제 해결 (getAuth → firebaseConfig의 auth 사용 통일
- Google/Kakao 로그인 시 Nest 서버 인증을 serverLogin()으로 통합
- auth.js에 Nest 서버 인증(serverLogin) 및 공통 로그아웃(logout) 로직 통합
- googleLogin.js, kakaoLogin.js는 인증 후 serverLogin 호출하도록 수정
- PrivateRoute에서 getAuth 대신 auth import 사용하도록 수정
- 전체 로그인 흐름 정리 및 로그 출력 일관화
