+++ client/src/components/PrivateRoute.jsx
@@ -1,16 +1,16 @@
-import { getAuth, onAuthStateChanged } from "firebase/auth";
-import { Navigate } from "react-router-dom";
 import { useEffect, useState } from "react";
+import { Navigate } from "react-router-dom";
+import { onAuthStateChanged } from "firebase/auth";
+import { auth } from "../utils/auth/firebaseConfig";
 
 export default function PrivateRoute({ children }) {
     const [user, setUser] = useState(undefined);
-    const auth = getAuth();
 
     useEffect(() => {
         const unsub = onAuthStateChanged(auth, (u) => setUser(u ?? null));
         return () => unsub();
-    }, [auth]);
+    }, []);
 
-    if (user === undefined) return null; // 로딩 중
+    if (user === undefined) return null;
     return user ? children : <Navigate to="/login" />;
 }
